+++ packages/shared/src/index.ts
@@ -1,4 +1,5 @@
 export * from './bookingUrl.js';
+export * from './bookingDraft.js';
 export * from './constants.js';
 export * from './discovery.js';
 export * from './roles.js';